backend/data_processing.py
```python
import pandas as pd
from sklearn.preprocessing import StandardScaler

# Define the schema for cosmological data
CosmologicalDataSchema = {
    'telescope_data': str,
    'satellite_data': str,
    'space_probe_data': str
}

# Function to process cosmological data
def processCosmologicalData(data):
    # Load the data into a pandas DataFrame
    df = pd.DataFrame(data, columns=CosmologicalDataSchema.keys())

    # Preprocess the data
    df = preprocessData(df)

    # Export the processed data
    cosmologicalData = df.to_dict(orient='records')

    return cosmologicalData

# Function to preprocess the data
def preprocessData(df):
    # Replace missing values with the median of the column
    df.fillna(df.median(), inplace=True)

    # Standardize the numerical columns
    scaler = StandardScaler()
    df[df.select_dtypes(include=['float64', 'int64']).columns] = scaler.fit_transform(df.select_dtypes(include=['float64', 'int64']))

    return df
```